Apply safe API cleanups and fix SQLite filtered count
API surface cleanups (from the interface review):
- Remove the duplicate ISearchableEntity interface; document the
  surviving ISearchableDocument.
- Finish the truncated MinScore doc comment and note that scores are
  not comparable across providers.
- Make SearchOptions.IncludeRawDocument init-only for consistency with
  the other options.
- Widen SearchResponse.TotalCount to long to match ISearchIndex.CountAsync,
  and read counts via GetInt64/Convert.ToInt64 in both providers.
- Give ISearchEngineManager.Get a default CancellationToken, matching the
  rest of the surface.
- Fix the README/nuget examples that referenced a non-existent MaxResults
  option (it is Take), and document Skip/IncludePartialMatches.

Bug fix surfaced by new tests:
- SQLite CountAsync(SearchRequest) ignored the full-text Query and only
  applied structured filters, disagreeing with Postgres and with
  SearchResponse.TotalCount. It now restricts the count to FTS matches
  when a query is present, mirroring SearchAsync.

+++ Source/SearchLite/Interfaces.cs
@@ -1,8 +1,18 @@
 ﻿namespace SearchLite;
 
+/// <summary>
+/// Indexable documents must implement this interface.
+/// </summary>
 public interface ISearchableDocument
 {
+    /// <summary>
+    /// Unique identifier within the index.
+    /// </summary>
     string Id { get; }
+
+    /// <summary>
+    /// The searchable text for the document.
+    /// </summary>
     string GetSearchText();
 }
 
@@ -18,7 +28,7 @@ public class SearchResult<T>
 public class SearchResponse<T>
 {
     public required IReadOnlyList<SearchResult<T>> Results { get; init; }
-    public int TotalCount { get; init; }
+    public long TotalCount { get; init; }
     public float MaxScore { get; init; }
     public TimeSpan SearchTime { get; init; }
 }
